use 64bit service handle
service handle is casted to uint32 which does not fit the 64bit pointers in 64 bit OS. Extend it to uint64.

+++ cspt_network_hub/service/inc/srv_common.h
@@ -8,7 +8,7 @@
 #ifndef SRV_COMMON_H_
 #define SRV_COMMON_H_
 
-typedef unsigned int service_handle;
+typedef unsigned long long int service_handle;
 
 #define SRV_INVALID 0xffffffff
